refactor: 更新多個彈出視窗的 CSS 引入，將 editServer.module.css 替換為 setting.module.css、applyFriend.module.css 替換為 apply.module.css，以提升樣式一致性和可讀性。
