+++ src/components/popups/CreateServer.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useRef, useState } from 'react';
 
 // CSS
 import popup from '@/styles/popup.module.css';
-import setting from '@/styles/popups/editServer.module.css';
+import setting from '@/styles/popups/setting.module.css';
 import createServer from '@/styles/popups/createServer.module.css';
 
 // Types
