+++ src/components/popups/Dialog.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useRef } from 'react';
 
 // CSS
 import popup from '@/styles/popup.module.css';
-import setting from '@/styles/popups/editServer.module.css';
+import setting from '@/styles/popups/setting.module.css';
 
 // Services
 import ipcService from '@/services/ipc.service';
