--- src/styles/popups/editChannel.module.css
@@ -1,12 +1,0 @@
-.body {
-  padding: 20px;
-  flex-direction: column;
-  justify-content: center;
-}
-
-.inputGroup {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  padding: 15px;
-}
