+++ src/styles/popups/setting.module.css
@@ -73,6 +73,17 @@
   padding-left: 20px;
 }
 
+.linkText {
+  color: #3b82f6;
+  transition: color 0.2s;
+}
+
+.linkText:hover {
+  color: #2563eb;
+  text-decoration: underline;
+  cursor: pointer;
+}
+
 /* Basic Info Section */
 
 .avatarWrapper {
@@ -254,4 +265,42 @@
   height: 4rem;
   resize: none;
   overflow-y: scroll;
+}
+
+.developerCardGrid {
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
+  gap: 1rem;
+  width: 100%;
+}
+
+.developerCard {
+  background-color: white;
+  border: 1px solid #e5e7eb;
+  border-radius: 0.5rem;
+  padding: 0.75rem;
+  transition: background-color 0.2s;
+}
+
+.developerCard:hover {
+  background-color: #f9fafb;
+}
+
+.developerCard .nameText {
+  color: #3b82f6;
+  transition: color 0.2s;
+  text-decoration: underline;
+  cursor: pointer;
+  display: block;
+  margin-bottom: 0.25rem;
+}
+
+.developerCard .nameText:hover {
+  color: #2563eb;
+}
+
+.developerCard .roleText {
+  color: #4b5563;
+  font-size: 0.9rem;
+  display: block;
 }
