--- src/styles/popups/systemSetting.module.css
@@ -1,48 +1,0 @@
-.link {
-  color: #3b82f6;
-  transition: color 0.2s;
-}
-
-.link:hover {
-  color: #2563eb;
-  text-decoration: underline;
-  cursor: pointer;
-}
-
-.developerGrid {
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  gap: 1rem;
-  width: 100%;
-}
-
-.developerCard {
-  background-color: white;
-  border: 1px solid #e5e7eb;
-  border-radius: 0.5rem;
-  padding: 0.75rem;
-  transition: background-color 0.2s;
-}
-
-.developerCard:hover {
-  background-color: #f9fafb;
-}
-
-.developerName {
-  color: #3b82f6;
-  transition: color 0.2s;
-  text-decoration: underline;
-  cursor: pointer;
-  display: block;
-  margin-bottom: 0.25rem;
-}
-
-.developerName:hover {
-  color: #2563eb;
-}
-
-.developerRole {
-  color: #4b5563;
-  font-size: 0.9rem;
-  display: block;
-}
